+++ lib/rblade/rails_template.rb
@@ -3,6 +3,7 @@ require "rblade/compiler"
 require "rblade/helpers/attributes_manager"
 require "rblade/helpers/class_manager"
 require "rblade/helpers/stack_manager"
+require "rblade/helpers/style_manager"
 
 module RBlade
   class RailsTemplate