+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## 0.1.1
 - Fix slots being calculated without parent's context
 - Add @props statement
-- Add attributes manager
+- Add attributes, class and style manager
 
 ## 0.1.0 [2024-07-22]
 - Initial release